requirements.txt
streamlit>=1.30
python-pptx>=0.6.21
requests>=2.31